Add files via upload
"chap8.R" serves as a comprehensive guide to understanding heteroskedasticity. It walks you through the basics, equips you with the skills to detect it through various methods (Lagrange multiplier tests, the White test, and the Goldfeld-Quandt test). Additionally, it provides insights into working with heteroskedasticity-consistent standard errors, showcasing the use of the 'sandwich' package and how to compute Heteroskedasticity-Consistent confidence intervals.

"chap9.R" delves into the application of time-series analysis focusing on stationary variables. It explores the concept of finite distributed lags and the significance of serial correlation, ACF correlograms, and formal tests for serially correlated errors like the Breusch-Godfrey and Durbin-Watson tests. Furthermore, it presents various models, such as Autoregressive Distributed Lag, and guides on choosing the best model using AIC and BIC. Lastly, it introduces forecasting techniques, including the use of exponential smoothing.

"chap10.R" begins with a straightforward review of the core concepts of unbiasedness, consistency, and efficiency, discussing what happens in cases where regressors and the error term are correlated. It provides a practical example of least squares estimation in wage equation. It then leads you through the process of using instrumental variables (IV) in the estimation, explains the challenges of weak instrumental variables, the workings of two-stage least squares (2SLS), and the importance of specification tests and testing instrument validity.

Lastly, "chap12.R" provides a discussion of time-series analysis with non-stationary variables. Starting with a comparison between stationary and non-stationary variables, it walks you through various models such as first-order autoregressive models with zero and non-zero means, as well as those that fluctuate around a linear trend. It also delves into the random walk models and introduces concepts of spurious regressions, unit root tests for stationarity, including the Dickey–Fuller and Augmented Dickey-Fuller tests, order of integration, and cointegration, ending with a look at the Error Correction Model.